pyproject.toml
[project]
name = "pyserver_users_allauth"
version = "0.0.1"
authors = [{ name = "Nynra", email = "nynradev@pm.me" }]
description = "A simple custom allauth users app for Django."
readme = { file = "README.txt", content-type = "text/markdown" }
requires-python = ">=3.10"
dependencies = [
    "Django>=4.0.0",
    "django-appconf>=1.0.4",
    # "django-rest-framework>=3.12.4",
    "django-allauth",
    "django-appconf",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: BSD License",
]
license = { file = "LICENSE" }

[build-system]
requires = ["setuptools>=61.0", "setuptools_scm>=8.0"]
build-backend = "setuptools.build_meta"
